+++ sdk-clients/tokens/tokens_types.gen.go
@@ -13,7 +13,7 @@ type BadRequestErrorDto struct {
 // ProviderTokenDto defines model for ProviderTokenDto.
 type ProviderTokenDto struct {
 	Address         string   `json:"address"`
-	ChainId         float32  `json:"chainId"`
+	ChainId         int      `json:"chainId"`
 	Decimals        float32  `json:"decimals"`
 	DisplayedSymbol string   `json:"displayedSymbol,omitempty"`
 	Eip2612         bool     `json:"eip2612,omitempty"`
@@ -34,7 +34,7 @@ type TagDto struct {
 // TokenDto defines model for TokenDto.
 type TokenDto struct {
 	Address   string   `json:"address"`
-	ChainId   float32  `json:"chainId"`
+	ChainId   int      `json:"chainId"`
 	Decimals  float32  `json:"decimals"`
 	Eip2612   bool     `json:"eip2612,omitempty"`
 	IsFoT     bool     `json:"isFoT,omitempty"`
@@ -49,7 +49,7 @@ type TokenDto struct {
 // TokenInfoDto defines model for TokenInfoDto.
 type TokenInfoDto struct {
 	Address    string                 `json:"address"`
-	ChainId    float32                `json:"chainId"`
+	ChainId    int                    `json:"chainId"`
 	Decimals   float32                `json:"decimals"`
 	Extensions map[string]interface{} `json:"extensions,omitempty"`
 	LogoURI    string                 `json:"logoURI"`